+++ src/components/BackToTop/BackToTop.scss
@@ -1,0 +1,31 @@
+.back-to-top {
+  position: fixed;
+  bottom: 15vh;
+  right: 80px;
+  border-bottom: none;
+  opacity: 0;
+  visibility: hidden;
+  transition: all 0.3s ease-in-out;
+
+  &--active {
+    opacity: 1;
+    visibility: visible;
+  }
+
+  &__icon {
+    border: 1px solid;
+    width: 25px;
+    height: 25px;
+    display: flex;
+    margin: 0 auto 8px;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+
+    &:before {
+      content: '\f3d8';
+      font-family: 'Ionicons';
+      font-style: normal;
+    }
+  }
+}
