+++ src/containers/Results/Results.js
@@ -3,44 +3,9 @@ import Result from './Result';
 import ResultFilter from './ResultFilter';
 import SearchInput from '../../components/SearchInput';
 import Pagination from '../../components/Pagination';
-import Button from '../../components/Button';
+import BackToTop from '../../components/BackToTop';
 import { getQueryParamObject } from '../../common/helpers';
 import './Results.scss';
-
-class BackToTop extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      activeClass: false
-    };
-  }
-  componentDidMount() {
-    window.addEventListener('scroll', () => {
-      if (window.pageYOffset >= window.innerHeight / 2) {
-        this.setState({
-          activeClass: true
-        });
-      } else {
-        this.setState({
-          activeClass: false
-        });
-      }
-    });
-  }
-  render() {
-    return (
-      <Button
-        buttonStyle="plain"
-        className={`back-to-top ${
-          this.state.activeClass ? 'back-to-top--active' : ''
-        }`}
-        onClick={() => window.scrollTo(0, 0)}
-      >
-        <i className="back-to-top__icon" />Back to Top
-      </Button>
-    );
-  }
-}
 
 class Results extends Component {
   componentDidMount() {
